♻️refactor: Rename fetchingActions -> initalizing...
