+++ src/components/Layout/Header/CourseSelector.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import Select from 'react-select';
 
-import { changeCoursebook } from '../../../actions/fetchingActions';
+import { changeCoursebook } from '../../../actions/initializingActions';
 import DropdownArrow from '../../Common/DropdownArrow.jsx';
 
 const mapStateToProps = ({ courseBook }) => {
@@ -14,7 +14,8 @@ const mapStateToProps = ({ courseBook }) => {
 };
 
 const mapDispatchToProps = dispatch => ({
-  selectCoursebook: ({ year, semester }) => dispatch(changeCoursebook({ year, semester })),
+  selectCoursebook: ({ year, semester }) =>
+    dispatch(changeCoursebook({ year, semester })),
 });
 
 // Utils for labels
@@ -22,16 +23,18 @@ const idxToString = [null, '1', 'S', '2', 'W'];
 const courseBookToString = b => `${b.year}-${idxToString[b.semester]}`;
 
 class CourseSelector extends Component {
-
-  selectCoursebook = selected => this.props.selectCoursebook(this.props.courseBooks[selected.value])
+  selectCoursebook = selected =>
+    this.props.selectCoursebook(this.props.courseBooks[selected.value]);
 
   render() {
     const { courseBooks, currentBook } = this.props;
     const courseBookOptions = courseBooks.map((book, i) => ({
-      value: i, label: courseBookToString(book), className: 'snutt__options',
+      value: i,
+      label: courseBookToString(book),
+      className: 'snutt__options',
     }));
-    const currentBookIndex = courseBooks.findIndex(b =>
-      b.semester === currentBook.semester && b.year === currentBook.year,
+    const currentBookIndex = courseBooks.findIndex(
+      b => b.semester === currentBook.semester && b.year === currentBook.year,
     );
     return (
       <div id="semester-wrapper">
@@ -50,4 +53,7 @@ class CourseSelector extends Component {
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(CourseSelector);
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(CourseSelector);
